backend/echo-chat/pyproject.toml
[project]
name = "echo-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastembed-gpu>=0.7.1",
    "google-adk>=1.5.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.14.3",
    "sentence-transformers>=5.0.0",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
[tool.ruff]
line-length = 100
target-version = "py313"
select = [
    "ALL",
]
ignore = [
    "PLR0913",
    "T201",
    "COM812",
    "DTZ001",

]
[tool.ruff.format]
skip-magic-trailing-comma = false
quote-style = 'single'
[tool.ruff.lint]
flake8-quotes = {inline-quotes = "single"}
